Refactor header_actions
